+++ pyaem/packagemanagerservicejson.py
@@ -3,6 +3,7 @@ from . import handlers
 import json
 import pycurl
 from . import result as res
+import urllib
 
 class PackageManagerServiceJson(object):
 
@@ -108,7 +109,7 @@ class PackageManagerServiceJson(object):
 
         method = 'post'
         url = '{0}/crx/packmgr/service/.json/etc/packages/{1}/{2}-{3}.zip'.format(
-            self.url, group_name, package_name, package_version)
+            self.url, group_name, urllib.quote(package_name), package_version)
         params = dict(params.items() + kwargs.items())
         _handlers = {
             201: _handler_failure