+++ test/test_packagemanagerservicejson.py
@@ -139,6 +139,19 @@ class TestPackageManagerServiceJson(unittest.TestCase):
             debug=True)
 
 
+    def test_install_package_spaces(self):
+
+        self.package_manager.install_package('mygroup', 'CQ 5.6.1 Security Service Pack', '1.2.3', foo='bar')
+        bag.request.assert_called_once_with(
+            'post',
+            'http://localhost:4502/crx/packmgr/service/.json/' +
+                'etc/packages/mygroup/CQ%205.6.1%20Security%20Service%20Pack-1.2.3.zip',
+            {'cmd': 'install',
+             'foo': 'bar'},
+            HandlersMatcher([200, 201, 401]),
+            debug=True)
+
+
     def test_replicate_package(self):
 
         self.package_manager.replicate_package('mygroup', 'mypackage', '1.2.3', foo='bar')